Move BuildCoordinationStatus to constants module
