+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/notifications/buildTask/BuildStatusMQNotifications.java
@@ -21,7 +21,7 @@ import org.jboss.pnc.messaging.spi.BuildStatusChanged;
 import org.jboss.pnc.messaging.spi.Message;
 import org.jboss.pnc.messaging.spi.MessageSender;
 import org.jboss.pnc.messaging.spi.Status;
-import org.jboss.pnc.spi.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.events.BuildCoordinationStatusChangedEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,6 +30,7 @@ import javax.enterprise.context.Dependent;
 import javax.enterprise.event.Observes;
 import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
