+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/ConfigurationsTest.java
@@ -23,7 +23,7 @@ import org.jboss.arquillian.junit.InSequence;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationSet;
 import org.jboss.pnc.model.User;
-import org.jboss.pnc.spi.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.BuildSetStatus;
 import org.jboss.pnc.spi.RebuildMode;
@@ -37,6 +37,7 @@ import org.junit.runner.RunWith;
 
 import javax.inject.Inject;
 import javax.persistence.PersistenceException;
+
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
