+++ integration-test/src/test/java/org/jboss/pnc/integration/remote/RemoteBuildTest.java
@@ -27,7 +27,7 @@ import org.jboss.pnc.model.RepositoryConfiguration;
 import org.jboss.pnc.rest.notifications.websockets.NotificationsEndpoint;
 import org.jboss.pnc.rest.restmodel.BuildConfigurationRest;
 import org.jboss.pnc.rest.restmodel.BuildRecordRest;
-import org.jboss.pnc.spi.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.RebuildMode;
 import org.jboss.pnc.spi.notifications.model.BuildChangedPayload;
