+++ rest/src/test/java/org/jboss/pnc/notifications/NotificationFactoryForBuildTest.java
@@ -18,7 +18,7 @@
 package org.jboss.pnc.notifications;
 
 import org.jboss.pnc.rest.notifications.DefaultNotificationFactory;
-import org.jboss.pnc.spi.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.coordinator.events.DefaultBuildStatusChangedEvent;
 import org.jboss.pnc.spi.events.BuildCoordinationStatusChangedEvent;
 import org.jboss.pnc.spi.notifications.model.BuildChangedPayload;
